feat: dockerize
