+++ openadapt/alembic/versions/98505a067995_add_browserevent_table.py
@@ -20,15 +20,15 @@ def upgrade() -> None:
     # ### commands auto generated by Alembic - please adjust! ###
     op.create_table('browser_event',
     sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('recording_timestamp', openadapt.models.ForceFloat(precision=10, scale=2, asdecimal=False), nullable=True),
+    sa.Column('recording_timestamp', openadapt.decorators.ForceFloat(precision=10, scale=2, asdecimal=False), nullable=True),
     sa.Column('recording_id', sa.Integer(), nullable=True),
     sa.Column('message', sa.JSON(), nullable=True),
-    sa.Column('timestamp', openadapt.models.ForceFloat(precision=10, scale=2, asdecimal=False), nullable=True),
+    sa.Column('timestamp', openadapt.decorators.ForceFloat(precision=10, scale=2, asdecimal=False), nullable=True),
     sa.ForeignKeyConstraint(['recording_id'], ['recording.id'], name=op.f('fk_browser_event_recording_id_recording')),
     sa.PrimaryKeyConstraint('id', name=op.f('pk_browser_event'))
     )
     with op.batch_alter_table('action_event', schema=None) as batch_op:
-        batch_op.add_column(sa.Column('browser_event_timestamp', openadapt.models.ForceFloat(precision=10, scale=2, asdecimal=False), nullable=True))
+        batch_op.add_column(sa.Column('browser_event_timestamp', openadapt.decorators.ForceFloat(precision=10, scale=2, asdecimal=False), nullable=True))
         batch_op.add_column(sa.Column('browser_event_id', sa.Integer(), nullable=True))
         batch_op.create_foreign_key(batch_op.f('fk_action_event_browser_event_id_browser_event'), 'browser_event', ['browser_event_id'], ['id'])
