+++ openadapt/models.py
@@ -21,26 +21,9 @@ from openadapt.drivers import anthropic
 from openadapt.db import db
 from openadapt.privacy.base import ScrubbingProvider, TextScrubbingMixin
 from openadapt.privacy.providers import ScrubProvider
+from openadapt.decorators import ForceFloat
 
 EMPTY_VALS = [None, "", [], (), {}]
-
-
-# https://groups.google.com/g/sqlalchemy/c/wlr7sShU6-k
-class ForceFloat(sa.TypeDecorator):
-    """Custom SQLAlchemy type decorator for floating-point numbers."""
-
-    impl = sa.Numeric(10, 2, asdecimal=False)
-    cache_ok = True
-
-    def process_result_value(
-        self,
-        value: int | float | str | None,
-        dialect: str,
-    ) -> float | None:
-        """Convert the result value to float."""
-        if value is not None:
-            value = float(value)
-        return value
 
 
 class Recording(db.Base):
